テスト用のアップデートレートを変えるとテストが落ちる問題を修正

+++ moorestech_server/Assets/Scripts/Tests/UnitTest/Core/Block/MinerSaveLoadTest.cs
@@ -49,7 +49,7 @@ namespace Tests.UnitTest.Core.Block
                     .GetField("_openableInventoryItemDataStoreService", BindingFlags.Instance | BindingFlags.NonPublic)
                     .GetValue(originalMinerComponent);
             var loadedRemainingMillSecond =
-                (int)typeof(VanillaElectricMinerComponent)
+                (double)typeof(VanillaElectricMinerComponent)
                     .GetField("_remainingMillSecond", BindingFlags.Instance | BindingFlags.NonPublic)
                     .GetValue(loadedMinerComponent);
